feat(android-recyclerview): remove unsubscribeSafely and update recycler adapter injection to call injector.unsubscribe directly
